ENH: allow title and wrapping control through jinja template
This adds a single file mode to allow creating a single file for
each release. It also allows to disable automatic indentation of
bullet point items, since not all projects necessarily use
bullet points for release notes. In single file mode, the output
file is a format string of the `name`, `version`, and `project_date`
and the file is fully replaced (`start_line` is not supported).

`name`, `version`, and `project_date` are now available in the main
template, allowing to render the title there. To make this useful,
the `title_format` should be set to `false`.